fix: register int64_t wrappers for scalar APIs to satisfy PyTorch type constraints; keep internal impls int

+++ csrc/utils/math.hpp
@@ -17,8 +17,8 @@ static constexpr T align(const T& a, const T& b) {
     return ceil_div(a, b) * b;
 }
 
-static int64_t get_tma_aligned_size(const int64_t& x, const int64_t& element_size) {
-    constexpr int64_t kNumTMAAlignmentBytes = 16;
+static int get_tma_aligned_size(const int& x, const int& element_size) {
+    constexpr int kNumTMAAlignmentBytes = 16;
     DG_HOST_ASSERT(kNumTMAAlignmentBytes % element_size == 0);
     return align(x, kNumTMAAlignmentBytes / element_size);
 }
